Add Orientation notifier

+++ application/app.go
@@ -67,12 +67,15 @@ func Orientation() layout.Edge {
 
 var orientationNotifier comm.IntValue
 
+// In order to receive orientation notification on android you may need to set
+// `android:configChanges="orientation|keyboardHidden|screenSize"` in your
+// activity's manifest (https://stackoverflow.com/a/6005624).
 func OrientationNotifier() comm.IntNotifier {
 	return &orientationNotifier
 }
 
 func init() {
-	bridge.RegisterFunc("gomatcha.io/matcha/application SetOrientation", func(edge int) {
-		orientationNotifier.SetValue(edge)
+	bridge.RegisterFunc("gomatcha.io/matcha/application SetOrientation", func(edge int64) {
+		orientationNotifier.SetValue(int(edge))
 	})
 }
